requirements.txt
pytest==7.4.3
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
allure-pytest==2.13.2
python-dotenv==1.0.0

